t_stitch_jittor_engine/requirements.txt
jittor>=1.3.10.0
numpy>=1.23
Pillow>=9.0

